Refactor into more idiomatic structure
Move all the functionality other than starting the web server out of
`main.rs` and into a `lib` crate.

This was necessary to be able to move the integration tests into their
own file since you can't import binary crates for use in a test.

Moved all the web server definition stuff into its own file for now. I
don't love this place since it's not grouped based on any sort of
domain concept, but it's alright for now.
