+++ Cargo.toml
@@ -1,8 +1,16 @@
 [package]
-name = "rs-readme"
+name = "rs_readme"
 version = "0.1.0"
 authors = ["Greg Cline <gregrcline@gmail.com>"]
 edition = "2018"
+
+[[bin]]
+name = "rs_readme_bin"
+path = "src/main.rs"
+
+[lib]
+name = "rs_readme"
+path = "src/lib.rs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -25,3 +33,4 @@ http-service-mock = "0.4.0"
 http = "0.1.19"
 http-service = "0.4.0"
 pretty_assertions = "0.6.1"
+
