+++ src/lib.rs
@@ -1,0 +1,13 @@
+#[macro_use]
+extern crate horrorshow;
+#[macro_use]
+extern crate serde_derive;
+
+mod content_finder;
+mod markdown_converter;
+mod static_files;
+mod web_server;
+
+pub use content_finder::{ContentError, ContentFinder, FileFinder};
+pub use markdown_converter::{Converter, MarkdownConverter, MarkdownError};
+pub use web_server::{build_app, State};
