+++ src/main.rs
@@ -1,158 +1,9 @@
-#[macro_use]
-extern crate serde_derive;
-#[macro_use]
-extern crate horrorshow;
-
-mod content_finder;
-mod markdown_converter;
-mod static_files;
-
-use horrorshow::helper::doctype;
-use horrorshow::prelude::*;
-use tide;
-use tide::{middleware, Request, Response, Server};
-
 use std::env;
 use std::path::PathBuf;
 
-use content_finder::{ContentError, ContentFinder, FileFinder};
-use markdown_converter::{Converter, MarkdownConverter};
-
-/// The state necessary to process requests.
-///
-/// It needs something to find some markdown content based on a URL path and something to take that
-/// markdown content and convert it to HTML to display.
-pub struct State<M, C>
-where
-    M: markdown_converter::MarkdownConverter,
-    C: content_finder::ContentFinder,
-{
-    markdown_converter: M,
-    content_finder: C,
-}
-
-/// The basic HTML of our page, the `<head>` and CSS and `<body>`
-fn base_html(converted: String) -> String {
-    format!(
-        "{}",
-        html! {
-            : doctype::HTML;
-            html {
-                head {
-                    link(rel="stylesheet", href="/static/octicons/octicons.css");
-                    link(rel="stylesheet", href="https://github.githubassets.com/assets/frameworks-146fab5ea30e8afac08dd11013bb4ee0.css");
-                    link(rel="stylesheet", href="https://github.githubassets.com/assets/site-897ad5fdbe32a5cd67af5d1bdc68a292.css");
-                    link(rel="stylesheet", href="https://github.githubassets.com/assets/github-c21b6bf71617eeeb67a56b0d48b5bb5c.css");
-                    title : "readme-rs";
-                }
-                body {
-                    : Raw(&converted);
-                }
-            }
-        }
-    )
-}
-
-/// The error HTML indicating the requested file is not markdown
-/// and therefore can't be rendered.
-fn not_markdown(file: &str) -> String {
-    format!(
-        "{}",
-        html! {
-            h1 : "Not a Markdown File";
-            p {
-                strong : file;
-                : " is not a markdown file and cannot be rendered";
-            }
-        }
-    )
-}
-
-/// The `tide::Endpoint` to render the `README.md`.
-///
-/// It assumes that there will be a `README.md` in your folder. It lets us have a special error
-/// message for it and lets the root of the website render `README.md`. It might not be necessary
-/// though, maybe we could just redirect `/` to `/README.md`.
-async fn render_readme(
-    req: Request<State<impl MarkdownConverter + Send + Sync, impl ContentFinder + Send + Sync>>,
-) -> tide::Result {
-    let state = req.state();
-
-    let contents = state
-        .content_finder
-        .content_for("README.md")
-        .map_err(|_| Response::new(404).body_string("Could not find README.md".to_string()))?;
-
-    let converted = state
-        .markdown_converter
-        .convert_markdown(&contents)
-        .await
-        .map_err(|_err| {
-            Response::new(500).body_string(format!(
-                "Could not convert the following markdown:\n {}",
-                &contents
-            ))
-        })?;
-
-    let resp = base_html(converted);
-
-    Ok(Response::new(200)
-        .body_string(resp)
-        .set_mime(mime::TEXT_HTML_UTF_8))
-}
-
-/// Renders any given file path containing markdown as HTML.
-async fn render_markdown_path(
-    req: Request<State<impl MarkdownConverter + Send + Sync, impl ContentFinder + Send + Sync>>,
-) -> tide::Result {
-    let state = req.state();
-
-    let contents = state
-        .content_finder
-        .content_for(&format!(".{}", req.uri().path()))
-        .map_err(|err| match err {
-            ContentError::NotMarkdown => Response::new(400)
-                .body_string(base_html(not_markdown(req.uri().path())))
-                .set_mime(mime::TEXT_HTML_UTF_8),
-            ContentError::CouldNotFetch => {
-                Response::new(404).body_string(format!("Could not find {}", req.uri().path()))
-            }
-        })?;
-
-    let converted = state
-        .markdown_converter
-        .convert_markdown(&contents)
-        .await
-        .map_err(|_| {
-            Response::new(500).body_string(format!(
-                "Could not convert the following markdown:\n {}",
-                &contents
-            ))
-        })?;
-
-    let resp = base_html(converted);
-
-    Ok(Response::new(200)
-        .body_string(resp)
-        .set_mime(mime::TEXT_HTML_UTF_8))
-}
-
-/// Builds a `tide::Server` with the appropriate endpoint mappings.
-fn build_app(
-    state: State<
-        impl MarkdownConverter + Send + Sync + 'static,
-        impl ContentFinder + Send + Sync + 'static,
-    >,
-) -> Server<State<impl MarkdownConverter, impl ContentFinder>> {
-    let mut app = Server::with_state(state);
-    app.middleware(middleware::RequestLogger::new());
-    app.at("").get(render_readme);
-    app.at("/static/octicons/:file")
-        .get(static_files::static_content);
-    app.at("/*").get(render_markdown_path);
-
-    app
-}
+use rs_readme::Converter;
+use rs_readme::FileFinder;
+use rs_readme::{build_app, State};
 
 #[async_std::main]
 async fn main() -> std::result::Result<(), std::io::Error> {
@@ -162,311 +13,12 @@ async fn main() -> std::result::Result<(), std::io::Error> {
         env::var("PORT").unwrap_or_else(|_| "4000".to_string())
     );
 
-    let state = State {
-        markdown_converter: Converter::new("https://api.github.com".to_string()),
-        content_finder: FileFinder::new(PathBuf::from(".")),
-    };
+    let state = State::new(
+        Converter::new("https://api.github.com".to_string()),
+        FileFinder::new(PathBuf::from(".")),
+    );
 
     let app = build_app(state);
 
     app.listen(addr).await
-}
-
-#[cfg(test)]
-mod test {
-    // Create mock
-    use super::*;
-    use async_std::io::ReadExt;
-    use async_trait::async_trait;
-    use content_finder::ContentError;
-    use http_service::Body;
-    use http_service_mock::make_server;
-    use markdown_converter::MarkdownError;
-    use pretty_assertions::assert_eq;
-    use std::collections::HashSet;
-    use std::sync::{Arc, Mutex};
-
-    /// A mock [`MarkdownConverter`] that returns:
-    /// `<h1>A Readme</h1>`
-    struct MockConverter;
-
-    #[async_trait]
-    impl MarkdownConverter for MockConverter {
-        async fn convert_markdown(&self, _md: &str) -> Result<String, MarkdownError> {
-            Ok("<h1>A Readme</h1>".to_string())
-        }
-    }
-
-    /// A mock [`ContentFinder`] the returns:
-    /// `# A Readme`
-    struct MockFinder;
-
-    impl ContentFinder for MockFinder {
-        fn content_for(&self, _resource: &str) -> Result<String, ContentError> {
-            Ok("# A Readme".to_string())
-        }
-    }
-
-    /// A mock [`ContentFinder`] and [`MarkdownConverter`] that keeps track of arguments
-    ///
-    /// Intended to be used to verify that an endpoint is calling its dependencies in
-    /// the expected way. It takes an `Arc<Mutex<HashSet<String>>>` so you can query
-    /// the `HashSet` later to verify what was placed in it.
-    ///
-    /// The `Arc` and `Mutex` are necessary for working across threads/async runtimes.
-    struct MockAssertSeen {
-        seen: Arc<Mutex<HashSet<String>>>,
-    }
-
-    impl MockAssertSeen {
-        fn new(seen: Arc<Mutex<HashSet<String>>>) -> MockAssertSeen {
-            MockAssertSeen { seen }
-        }
-    }
-
-    impl ContentFinder for MockAssertSeen {
-        fn content_for(&self, resource: &str) -> Result<String, ContentError> {
-            self.seen
-                .lock()
-                .expect("Could not lock mutex in content_for")
-                .insert(resource.to_string());
-
-            Ok(format!("content for: {}", resource).to_string())
-        }
-    }
-
-    #[async_trait]
-    impl MarkdownConverter for MockAssertSeen {
-        async fn convert_markdown(&self, md: &str) -> Result<String, MarkdownError> {
-            self.seen
-                .lock()
-                .expect("Could not lock mutex in convert_markdown")
-                .insert(md.to_string());
-
-            Ok(md.to_string())
-        }
-    }
-
-    #[async_std::test]
-    async fn index_wraps_in_html() {
-        // Setup
-        let state = State {
-            markdown_converter: MockConverter,
-            content_finder: MockFinder,
-        };
-        let app = build_app(state);
-        let mut server = make_server(app.into_http_service()).unwrap();
-
-        // Request
-        let req = http::Request::get("/").body(Body::empty()).unwrap();
-        let res = server.simulate(req).unwrap();
-
-        // Assert
-        let status = res.status();
-        assert_eq!(status.as_u16(), 200);
-
-        // End the borrow of res so we can consume it for the body
-        {
-            let mime = res
-                .headers()
-                .get("content-type")
-                .expect("Could not get content-type");
-            assert_eq!(mime, "text/html; charset=utf-8");
-        }
-
-        let mut body = String::with_capacity(1);
-        res.into_body().read_to_string(&mut body).await.unwrap();
-        let expected_body = "\
-<!DOCTYPE html>\
-<html>\
-  <head>\
-  <link rel=\"stylesheet\" href=\"/static/octicons/octicons.css\">\
-  <link rel=\"stylesheet\" href=\"https://github.githubassets.com/assets/frameworks-146fab5ea30e8afac08dd11013bb4ee0.css\">\
-  <link rel=\"stylesheet\" href=\"https://github.githubassets.com/assets/site-897ad5fdbe32a5cd67af5d1bdc68a292.css\">\
-  <link rel=\"stylesheet\" href=\"https://github.githubassets.com/assets/github-c21b6bf71617eeeb67a56b0d48b5bb5c.css\">\
-    <title>readme-rs</title>\
-  </head>\
-  <body>\
-    <h1>A Readme</h1>\
-  </body>\
-</html>";
-        assert_eq!(body, expected_body);
-    }
-
-    #[async_std::test]
-    async fn calls_content_finder_with_file_path() {
-        // Setup
-        let converter = Arc::new(Mutex::new(HashSet::new()));
-        let finder = Arc::new(Mutex::new(HashSet::new()));
-        let state = State {
-            markdown_converter: MockAssertSeen::new(converter.clone()),
-            content_finder: MockAssertSeen::new(finder.clone()),
-        };
-        let app = build_app(state);
-        let mut server = make_server(app.into_http_service()).unwrap();
-
-        // Request
-        let req = http::Request::get("/test_dir/a.md")
-            .body(Body::empty())
-            .unwrap();
-        let res = server.simulate(req).unwrap();
-
-        // Assert
-        let status = res.status();
-        assert_eq!(status.as_u16(), 200);
-
-        // End the borrow of res so we can consume it for the body
-        {
-            let mime = res
-                .headers()
-                .get("content-type")
-                .expect("Could not get content-type");
-            assert_eq!(mime, "text/html; charset=utf-8");
-        }
-
-        assert!(finder
-            .lock()
-            .expect("Could not lock in finder assert")
-            .contains("./test_dir/a.md"));
-        assert!(converter
-            .lock()
-            .expect("Could not lock in converter assert")
-            .contains("content for: ./test_dir/a.md"));
-    }
-
-    #[async_std::test]
-    async fn returns_400_for_non_md_file() {
-        // Create mock
-        struct MockFinderError;
-
-        impl ContentFinder for MockFinderError {
-            fn content_for(&self, _resource: &str) -> Result<String, ContentError> {
-                Err(ContentError::NotMarkdown)
-            }
-        }
-
-        // Setup
-        let state = State {
-            markdown_converter: MockConverter,
-            content_finder: MockFinderError,
-        };
-        let app = build_app(state);
-        let mut server = make_server(app.into_http_service()).unwrap();
-
-        // Request
-        let req = http::Request::get("/foo.txt").body(Body::empty()).unwrap();
-        let res = server.simulate(req).unwrap();
-
-        // Assert
-        let status = res.status();
-        assert_eq!(status.as_u16(), 400);
-
-        // End the borrow of res so we can consume it for the body
-        {
-            let mime = res
-                .headers()
-                .get("content-type")
-                .expect("Could not get content-type");
-            assert_eq!(mime, "text/html; charset=utf-8");
-        }
-
-        let mut body = String::with_capacity(1);
-        res.into_body().read_to_string(&mut body).await.unwrap();
-        let expected_body = "\
-<!DOCTYPE html>\
-<html>\
-  <head>\
-  <link rel=\"stylesheet\" href=\"/static/octicons/octicons.css\">\
-  <link rel=\"stylesheet\" href=\"https://github.githubassets.com/assets/frameworks-146fab5ea30e8afac08dd11013bb4ee0.css\">\
-  <link rel=\"stylesheet\" href=\"https://github.githubassets.com/assets/site-897ad5fdbe32a5cd67af5d1bdc68a292.css\">\
-  <link rel=\"stylesheet\" href=\"https://github.githubassets.com/assets/github-c21b6bf71617eeeb67a56b0d48b5bb5c.css\">\
-    <title>readme-rs</title>\
-  </head>\
-  <body>\
-    <h1>Not a Markdown File</h1>\
-    <p><strong>/foo.txt</strong> is not a markdown file and cannot be rendered</p>\
-  </body>\
-</html>";
-        assert_eq!(body, expected_body);
-    }
-
-    #[async_std::test]
-    async fn static_content_returns_appropriate_files() {
-        // Setup
-        let state = State {
-            markdown_converter: MockConverter,
-            content_finder: MockFinder,
-        };
-        let app = build_app(state);
-        let mut server = make_server(app.into_http_service()).unwrap();
-
-        // Expected results
-        // (path, status, mime, body)
-        let expected = vec![
-            (
-                "/static/octicons/octicons.css",
-                200,
-                "text/css; charset=utf-8",
-                {
-                    let mut vec = Vec::new();
-                    vec.extend_from_slice(include_bytes!("../static/octicons/octicons.css"));
-                    vec
-                },
-            ),
-            (
-                "/static/octicons/octicons.eot",
-                200,
-                "application/vnd.ms-fontobject",
-                {
-                    let mut vec = Vec::new();
-                    vec.extend_from_slice(include_bytes!("../static/octicons/octicons.eot"));
-                    vec
-                },
-            ),
-            ("/static/octicons/octicons.svg", 200, "image/svg+xml", {
-                let mut vec = Vec::new();
-                vec.extend_from_slice(include_bytes!("../static/octicons/octicons.svg"));
-                vec
-            }),
-            ("/static/octicons/octicons.ttf", 200, "font/ttf", {
-                let mut vec = Vec::new();
-                vec.extend_from_slice(include_bytes!("../static/octicons/octicons.ttf"));
-                vec
-            }),
-            ("/static/octicons/octicons.woff", 200, "font/woff", {
-                let mut vec = Vec::new();
-                vec.extend_from_slice(include_bytes!("../static/octicons/octicons.woff"));
-                vec
-            }),
-            ("/static/octicons/octicons.woff2", 200, "font/woff2", {
-                let mut vec = Vec::new();
-                vec.extend_from_slice(include_bytes!("../static/octicons/octicons.woff2"));
-                vec
-            }),
-        ];
-
-        for (path, status, mime, body) in expected.iter() {
-            // Make request
-            let req = http::Request::get(*path).body(Body::empty()).unwrap();
-            let res = server.simulate(req).unwrap();
-
-            // Assert
-            let res_status = res.status();
-            assert_eq!(&res_status.as_u16(), status, "path: {}", path);
-
-            // End the borrow of res so we can consume it for the body
-            {
-                let res_mime = res
-                    .headers()
-                    .get("content-type")
-                    .expect("Could not get content-type");
-                assert_eq!(res_mime, mime, "path: {}", path);
-            }
-
-            let mut res_body: Vec<u8> = Vec::with_capacity(1);
-            res.into_body().read_to_end(&mut res_body).await.unwrap();
-
-            assert_eq!(&res_body, body, "path: {}", path);
-        }
-    }
 }
